ETH_DIR: Fix the build for split repos
When using split repos ETH_DIR has not been set yet but is nonetheless
used in the eth_add_resources() call.

Any ETH_USE() call would set ETH_DIR but it seems that
eth_add_resources() has to be called before setting of the executable.

For umbrella project this is not a problem since before alethzero other
projects were already built and as such had the ETH_DIR set.